requirements.txt
openai>=1.0
anthropic>=0.18
lark-oapi>=1.0
pyyaml>=6.0
